pyproject.toml
[project]
name = "g"
version = "0.1.0"
description = "g stuff"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "glfw>=2.9.0",
  "numpy>=2.3.2",
  "pillow>=11.3.0",
  "pyopengl>=3.1.10",
  "rich>=14.1.0",
]

[tool.basedpyright]
reportExplicitAny = false
reportAny = false
reportImplicitStringConcatenation = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnusedCallResult = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["g_utils*", "g_game*"]
